// src/lib/database.types.ts
export type Database = {
  public: {
    Tables: {
      vehicles: {
        Row: {
          created_at: string
          description: string | null
          engine: string | null
          features: string[] | null
          fuel_type: string | null
          id: number
          image_urls: string[] | null
          make: string | null
          mileage: number | null
          model: string | null
          price: number | null
          transmission: string | null
          year: number | null
        }
        Insert: {
          created_at?: string
          description?: string | null
          engine?: string | null
          features?: string[] | null
          fuel_type?: string | null
          id?: number
          image_urls?: string[] | null
          make?: string | null
          mileage?: number | null
          model?: string | null
          price?: number | null
          transmission?: string | null
          year?: number | null
        }
        Update: {
          created_at?: string
          description?: string | null
          engine?: string | null
          features?: string[] | null
          fuel_type?: string | null
          id?: number
          image_urls?: string[] | null
          make?: string | null
          mileage?: number | null
          model?: string | null
          price?: number | null
          transmission?: string | null
          year?: number | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (Database["public"]["Tables"] & Database["public"]["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (Database["public"]["Tables"] &
        Database["public"]["Views"])
    ? (Database["public"]["Tables"] &
        Database["public"]["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof Database["public"]["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof Database["public"]["Enums"]
    ? Database["public"]["Enums"][PublicEnumNameOrOptions]
    : never